Renamed Balances to Assets in classes and string assets
