+++ app/src/main/java/io/gnosis/safe/ui/safe/assets/coins/CoinBalanceAdapter.kt
@@ -1,4 +1,4 @@
-package io.gnosis.safe.ui.safe.balances.coins
+package io.gnosis.safe.ui.safe.assets.coins
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
