+++ app/src/main/res/values/strings.xml
@@ -34,7 +34,7 @@
     <string name="agree">Agree</string>
     <string name="add_safe">Add Safe</string>
     <string name="switch_safes">Switch Safes</string>
-    <string name="balances">Balances</string>
+    <string name="assets">Assets</string>
     <string name="transactions">Transactions</string>
     <string name="settigns">Settings</string>
     <string name="remove_safe">Remove safe</string>
